Remove TestTelegramBot Command and Refactor OrderFactory Methods for Enhanced Order Creation. Deleted the unused TestTelegramBot command to streamline the codebase. Updated methods in OrderFactory to return structured data, allowing for better handling of order and item creation, including improved error management and data consistency. Adjusted the quantity field in the order items migration to use decimal values for better precision.

+++ database/migrations/2025_05_04_213628_create_order_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('order_id')->constrained()->onDelete('cascade');
             $table->morphs('item');
-            $table->integer('quantity')->default(1);
+            $table->decimal('quantity', 8, 2)->default(1);
             $table->decimal('unit_price', 10, 2);
             $table->timestamps();
